Implemented ando changes
