+++ app/assets/stylesheets/admin_interface.css.scss
@@ -1,0 +1,9 @@
+.admin-interface {
+  padding-left: 300px;
+}
+
+@media only screen and (max-width: 992px) {
+  .admin-interface {
+    padding-left: 0;
+  }
+}
