+++ app/assets/stylesheets/questionnaire.scss
@@ -17,13 +17,10 @@
   margin-left: -8px;
   padding-left: 6px;
 }
-.otherwise-textfield {
-  label {
-    margin-bottom: 20px;
-  }
-}
+
 .input-field.inline {
   margin-top: -21px;
+  margin-bottom: 0px;
   top: 16px;
 }
 .extra-spacing.row {
