+++ app/generators/date_generator.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-class DateGenerator < Generator
+class DateGenerator < QuestionTypeGenerator
+  DATEFIELD_PLACEHOLDER = 'Vul een datum in'
+
   def generate(question)
     title = safe_join([question[:title].html_safe, generate_tooltip(question[:tooltip])])
     safe_join([content_tag(:p, title, class: 'flow-text'), mydate_field(question)])
