+++ app/generators/questionnaire_question_generator.rb
@@ -23,21 +23,32 @@ class QuestionnaireQuestionGenerator < Generator
     question_body = find_generator(question[:type]).generate(question)
     question_body = content_tag(:div, question_body, class: 'col s12')
     question_body = content_tag(:div, question_body,
-                                class: find_generator(:klasses).generate(question).to_s)
+                                class: find_generator(:klasses, only_questions: false).generate(question).to_s)
     wrap_question_in_sections(question_body, question)
   end
 
-  def find_generator(type)
+  private
+
+  def find_generator(type, only_questions: true)
     generator = @generators[type]
-    return generator if generator.present?
-    raise "Unknown question type #{type}"
+    check_question_type_available(type, generator)
+    check_question_type_allowed(type, generator, only_questions)
+    generator
   end
 
   def wrap_question_in_sections(question_body, question)
     body = []
-    body << find_generator(:section_start).generate(question)
+    body << find_generator(:section_start, only_questions: false).generate(question)
     body << question_body
-    body << find_generator(:section_end).generate(question)
+    body << find_generator(:section_end, only_questions: false).generate(question)
     body.compact
+  end
+
+  def check_question_type_available(type, generator)
+    raise "Unknown question type #{type}" unless generator.present?
+  end
+
+  def check_question_type_allowed(type, generator, only_questions)
+    raise "Question type #{type} not allowed as question" if only_questions && !generator.is_a?(QuestionTypeGenerator)
   end
 end
