+++ app/generators/raw_generator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class RawGenerator < Generator
+class RawGenerator < QuestionTypeGenerator
   def generate(question)
     question[:content].html_safe
   end
