+++ app/generators/textarea_generator.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-class TextareaGenerator < Generator
+class TextareaGenerator < QuestionTypeGenerator
+  TEXTAREA_PLACEHOLDER = 'Vul iets in'
+
   def generate(question)
     title = safe_join([question[:title].html_safe, generate_tooltip(question[:tooltip])])
     safe_join([content_tag(:p, title, class: 'flow-text'), textarea_field(question)])
