+++ app/generators/textfield_generator.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-class TextfieldGenerator < Generator
+class TextfieldGenerator < QuestionTypeGenerator
+  TEXTFIELD_PLACEHOLDER = 'Vul iets in'
+
   def generate(question)
     title = safe_join([question[:title].html_safe, generate_tooltip(question[:tooltip])])
     safe_join([content_tag(:p, title, class: 'flow-text'), textfield_field(question)])
