+++ app/views/admin/export.html.haml
@@ -59,9 +59,7 @@
                     admin_proof_of_participation_export_path(format: 'csv'),
                     class: 'waves-effect waves-light btn download-button')
 
-              - @pilot_questionnaires = Questionnaire.pilot
-              - @rest_questionnaires = Questionnaire.all - @pilot_questionnaires
-              - @rest_questionnaires.each do |questionnaire|
+              - @normal_questionnaires.each do |questionnaire|
                 %li.collection-item.avatar{:style => "min-height:65px"}
                   %i.large.material-icons.circle.blue.lighten-2 description
                   %span.title
