+++ app/views/admin/preview_overview.html.haml
@@ -4,5 +4,5 @@
     %h1 Preview questionnaires
     .section
       = form_tag('/admin/preview', method: 'post') do
-        = select_tag(:id, options_for_select(questionnaire_select_options(@used_questionnaires.sort_by{|x|x.name}), disabled: '', selected: ''), required: true)
+        = select_tag(:id, options_for_select(questionnaire_select_options(@normal_questionnaires.sort_by{|x|x.name}), disabled: '', selected: ''), required: true)
         = button_tag('Preview questionnaire', type: 'submit', class: 'btn waves-effect waves-light')
