+++ app/views/questionnaire/informed_consent.html.haml
@@ -4,5 +4,5 @@
   title: @protocol.informed_consent_questionnaire.title,
   submit_text: 'Volgende',
   action: '/informed_consent',
-  unsubscribe_url: Rails.application.routes.url_helpers.questionnaire_path(uuid: @response.uuid),
+  unsubscribe_url: @response.unsubscribe_url,
   params: { authenticity_token: form_authenticity_token(form_options: { action: '/informed_consent', method: 'post' })})
